pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.2",
    "fastapi>=0.115.12",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "plaid-python>=29.1.0",
    "plotly>=6.0.1",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.1",
    "uvicorn>=0.34.0",
]
